Changed some code so the squares can move

+++ src/gfx/renderer.h
@@ -5,8 +5,8 @@
 
 struct quad
 {
-    i32 x, y;
-    u32 w, h;
+    f32 x, y;
+    f32 w, h;
     f32 rotation;
 };
